Create SeleniumFileDownload
If a webpage is loaded with tons of JS including buttons for downloads, then scraping using BeautifulSoup will not be possible. The code allows for the download of a file be simulating a click on the JS button. You will however, first have to find the <a class=...</a> tag and provide the class value. Obviously this code only makes sense to automate a download from a site and can't be used dynamically as this user input is required.